backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
psutil>=5.9
bcrypt>=4.1
python-dotenv>=1.0

# Tests
pytest>=8.0
httpx>=0.26

